# swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_healthCheck",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/minuta": {
      "post": {
        "operationId": "MinutaController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMinutaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Minuta"
        ]
      },
      "get": {
        "operationId": "MinutaController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Minuta"
        ]
      }
    },
    "/minuta/{id}": {
      "get": {
        "operationId": "MinutaController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Minuta"
        ]
      },
      "patch": {
        "operationId": "MinutaController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMinutaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Minuta"
        ]
      },
      "delete": {
        "operationId": "MinutaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Minuta"
        ]
      }
    }
  },
  "info": {
    "title": "Minuta Contractual MID",
    "description": "API",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "API",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "CreateMinutaDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateMinutaDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}